# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NovelHub",
    "version": "0.1.0"
  },
  "paths": {
    "/api/novels/": {
      "get": {
        "tags": [
          "novels"
        ],
        "summary": "Get Novels",
        "operationId": "get_novels_api_novels__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "novels"
        ],
        "summary": "Upload Novel",
        "operationId": "upload_novel_api_novels__post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_novel_api_novels__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}": {
      "get": {
        "tags": [
          "novels"
        ],
        "summary": "Get Novel By Id",
        "operationId": "get_novel_by_id_api_novels__novel_id__get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/wiki/generate": {
      "post": {
        "tags": [
          "wiki"
        ],
        "summary": "Generate Wiki Endpoint",
        "operationId": "generate_wiki_endpoint_api_wiki_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WikiGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/wiki/pages": {
      "get": {
        "tags": [
          "wiki"
        ],
        "summary": "List Wiki Pages Endpoint",
        "operationId": "list_wiki_pages_endpoint_api_wiki_pages_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiPageList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/wiki/pages/{page_id}": {
      "get": {
        "tags": [
          "wiki"
        ],
        "summary": "Get Wiki Page Endpoint",
        "operationId": "get_wiki_page_endpoint_api_wiki_pages__page_id__get",
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "wiki"
        ],
        "summary": "Delete Wiki Page Endpoint",
        "operationId": "delete_wiki_page_endpoint_api_wiki_pages__page_id__delete",
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Wiki Page Endpoint Api Wiki Pages  Page Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_novel_api_novels__post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "author": {
            "type": "string",
            "title": "Author",
            "default": ""
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_novel_api_novels__post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NovelList": {
        "properties": {
          "novels": {
            "items": {
              "$ref": "#/components/schemas/NovelResponse"
            },
            "type": "array",
            "title": "Novels"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "novels",
          "total"
        ],
        "title": "NovelList"
      },
      "NovelResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "file_hash": {
            "type": "string",
            "title": "File Hash"
          },
          "chapter_count": {
            "type": "integer",
            "title": "Chapter Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "author",
          "language",
          "file_hash"
        ],
        "title": "NovelResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WikiGenerateRequest": {
        "properties": {
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version",
            "default": "v1"
          }
        },
        "type": "object",
        "required": [
          "novel_id"
        ],
        "title": "WikiGenerateRequest"
      },
      "WikiPageList": {
        "properties": {
          "pages": {
            "items": {
              "$ref": "#/components/schemas/WikiPageResponse"
            },
            "type": "array",
            "title": "Pages"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "pages",
          "total"
        ],
        "title": "WikiPageList"
      },
      "WikiPageResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "source_chapters": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Chapters"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "novel_id",
          "entity_id",
          "title",
          "content",
          "language",
          "version",
          "source_chapters",
          "prompt_version",
          "is_published",
          "created_at",
          "updated_at"
        ],
        "title": "WikiPageResponse"
      }
    }
  }
}